HYDRA-FREE/requirements.txt
click==8.1.3
httpx==0.24.1
beautifulsoup4==4.12.2
pyyaml==6.0
playwright==1.35.0
fastapi==0.100.0
uvicorn==0.23.0
textblob==0.18.0.post0
python-dotenv==1.0.1

